Refactor to enable astronomy layer
