+++ ShipsClock/ClockFace.swift
@@ -11,8 +11,6 @@ import SwiftUI
 struct ClockFace: View {
     @EnvironmentObject var shipsClock: ShipsClock
 
-    let borderColor = Color.black
-    
     func diameter(_ geometry : GeometryProxy) -> CGFloat {
         CGFloat.minimum(geometry.size.width, geometry.size.height)
     }
@@ -41,184 +39,9 @@ struct ClockFace: View {
     var body: some View {
         GeometryReader { geometry in
             ZStack {
-                self.border
-                Numbers(radius: self.radius(geometry))
-                Ticks(radius: self.radius(geometry))
-                MinuteHand(
-                    radius: self.radius(geometry), timeInSeconds: self.shipsClock.timeOfDayInSeconds)
-                HourHand(
-                    radius: self.radius(geometry), timeInSeconds: self.shipsClock.timeOfDayInSeconds)
-                WatchHand(
-                    radius: self.radius(geometry), timeInSeconds: self.shipsClock.timeOfDayInSeconds)
-                CenterCircle(radius: self.radius(geometry))
+                ClockBackground(radius: self.radius(geometry))
+                ClockHands(radius: self.radius(geometry)).environmentObject(self.shipsClock)
             }.frame(width: self.diameter(geometry), height: self.diameter(geometry), alignment: .top)
-        }
-    }
-
-    var border: some View {
-        Circle().stroke(borderColor, lineWidth: 4)
-    }
-    
-    struct HourTicks: View {
-        var tickColor = Color.black
-        
-        var radius : Double
-        var hourInterval : Int
-        var widthMultiplier : Int
-        var tickOutside : Double
-        var tickInside : Double
-
-        var body: some View {
-            let baseLineWidth = CGFloat(radius / 60.0)
-            return Path { path in
-                for hour in stride(from: 0, to: 23, by: hourInterval) {
-                    let angle = ClockFace.hourAngle(forHour: hour)
-                    path.move(to: ClockFace.pointOnRadius(forAngle: angle, givenRadius: radius, atPosition: self.tickOutside))
-                    path.addLine(to: ClockFace.pointOnRadius(forAngle: angle, givenRadius: radius, atPosition: self.tickInside))
-                }
-            }
-            .stroke(lineWidth: CGFloat(widthMultiplier) * baseLineWidth)
-            .fill(self.tickColor)
-        }
-    }
-
-    struct Ticks: View {
-        var radius: Double
-        let watchTicksOutside = 0.8
-        let watchTicksInside = 0.7
-        let watchesTicksOutside = 0.68
-        let watchesTicksInside = 0.6
-
-        var body: some View {
-            ZStack {
-                HourTicks(radius: radius, hourInterval: 4, widthMultiplier: 1, tickOutside: watchesTicksOutside, tickInside: watchesTicksInside)
-                HourTicks(radius: radius, hourInterval: 6, widthMultiplier: 2, tickOutside: watchTicksOutside, tickInside: watchTicksInside)
-            }
-        }
-    }
-        
-    struct PositionedHourLabel: View {
-        let textRadiusMultiplier = 0.9
-        let hour: Int
-        let radius: Double
-        let angle: Double
-        
-        var body: some View {
-            Text(hour.description)
-                .position(ClockFace.pointOnRadius(forAngle: angle, givenRadius: radius, atPosition: textRadiusMultiplier))
-        }
-    }
-    
-    struct Numbers: View {
-        let hourCount = 24
-        let radius: Double
-        
-        var body: some View {
-            return ForEach(0..<self.hourCount, content: { hour in
-                PositionedHourLabel(
-                    hour: hour,
-                    radius: self.radius,
-                    angle: ClockFace.hourAngle(forHour: hour)
-                ).font(.system(size: CGFloat(self.radius / 9.0), weight: .black, design: .default))
-            })
-        }
-    }
-
-    struct HourHand: View {
-        var radius: Double
-        var timeInSeconds: Int
-        
-        let hourHandStart = -0.1
-        let hourHandEnd = 0.55
-        
-        var body: some View {
-            Path() { path in
-                let lengthOfDayInSeconds = 24 * 60 * 60
-                let partOfDay = Double(timeInSeconds) / Double(lengthOfDayInSeconds)
-                let hourAngle = 3.0 * Double.pi / 2.0 - partOfDay * Double.pi * 2.0
-                path.move(to: ClockFace.pointOnRadius(forAngle: hourAngle, givenRadius: radius, atPosition: hourHandStart))
-                path.addLine(to: ClockFace.pointOnRadius(forAngle: hourAngle, givenRadius: radius, atPosition: hourHandEnd))
-            }.stroke(lineWidth: CGFloat(radius / 22.0))
-        }
-    }
-    
-    struct WatchWand: View {
-        let transform: CGAffineTransform
-        
-        init(radius: Double, watchAngle: Double, atPosition watchHandEnd: Double) {
-            let center = ClockFace.pointOnRadius(forAngle: watchAngle, givenRadius: radius, atPosition: watchHandEnd)
-            let scale = CGFloat(radius / 90.0)
-            transform = CGAffineTransform.identity
-                .concatenating(CGAffineTransform.identity.scaledBy(x: scale, y: scale))
-                .concatenating(CGAffineTransform.identity.rotated(by: CGFloat(-watchAngle)))
-                .concatenating(CGAffineTransform.identity.translatedBy(x: center.x, y: center.y))
-        }
-
-        var body: some View {
-            Path() { path in
-                path.move(to: CGPoint(x: 0, y: 3))
-                path.addLine(to: CGPoint(x: -3, y: 0))
-                path.addLine(to: CGPoint(x: 0, y: -3))
-                path.addLine(to: CGPoint(x: 0, y: 3))
-            }.transform(transform).fill(Color.black)
-        }
-    }
-    
-    struct WatchHand: View {
-        let watchHandStart = -0.1
-        let watchHandEnd = 0.68
-
-        let lengthOfWatchInSeconds = 4 * 60 * 60
-        let radius: Double
-        let watchAngle: Double
-
-        init(radius: Double, timeInSeconds: Int) {
-            self.radius = radius
-            let watchRemainder = timeInSeconds % lengthOfWatchInSeconds
-            let partOfWatch = Double(watchRemainder) / Double(lengthOfWatchInSeconds)
-            watchAngle = Double.pi / 2.0 - partOfWatch * Double.pi * 2.0
-        }
-        
-        var body: some View {
-            ZStack {
-                Path() { path in
-                    path.move(to: ClockFace.pointOnRadius(forAngle: watchAngle, givenRadius: radius, atPosition: watchHandStart))
-                    path.addLine(to: ClockFace.pointOnRadius(forAngle: watchAngle, givenRadius: radius, atPosition: watchHandEnd))
-                }.stroke(lineWidth: CGFloat(radius / 35.0))
-                WatchWand(radius: radius, watchAngle: watchAngle, atPosition: watchHandEnd)
-            }
-        }
-    }
-    
-    struct MinuteHand: View {
-        var radius: Double
-        var timeInSeconds: Int
-        
-        let minuteHandStart = -0.1
-        let minuteHandEnd = 0.8
-        
-        var body: some View {
-            Path() { path in
-                let lengthOfHourInSeconds = 60 * 60
-                let hourRemainder = timeInSeconds % lengthOfHourInSeconds
-                let partOfHour = Double(hourRemainder) / Double(lengthOfHourInSeconds)
-                let minuteAngle = Double.pi / 2.0 - partOfHour * Double.pi * 2.0
-                path.move(to: ClockFace.pointOnRadius(forAngle: minuteAngle, givenRadius: radius, atPosition: minuteHandStart))
-                path.addLine(to: ClockFace.pointOnRadius(forAngle: minuteAngle, givenRadius: radius, atPosition: minuteHandEnd))
-            }.stroke(lineWidth: CGFloat(radius / 22.0))
-        }
-    }
-    
-    struct CenterCircle: View {
-        var radius: Double
-        
-        var body: some View {
-            let centerSize = CGFloat(radius / 45.0)
-            let offset = CGFloat(radius) - centerSize / 2.0
-            return Circle()
-                .size(width: centerSize, height: centerSize)
-                .offset(x: offset, y: offset)
-                .fill(Color.white)
         }
     }
 }
